+++ src/main/kotlin/blackjack/domain/participant/Participant.kt
@@ -4,16 +4,40 @@ import blackjack.domain.card.Card
 import blackjack.domain.card.Hand
 import blackjack.domain.card.Point
 
-interface Participant {
-    val name: String
-    val hand: Hand
-    val state: State
+abstract class Participant(
+    val name: String,
+    val hand: Hand,
+    state: State = Hittable
+) {
+    var state: State = state
+        protected set
 
     fun point(): Point = hand.calculate()
-    fun isPlayable(askHit: (String) -> Boolean): Boolean
-    fun receive(card: Card)
-    fun open(): Hand = hand
-    fun match(dealer: Dealer): Match
+
+    abstract fun isPlayable(askHit: (String) -> Boolean): Boolean
+
+    fun receive(card: Card) {
+        check(isPlayableState()) { "카드를 더 이상 받을 수 없는 상태입니다." }
+        hand.add(card)
+        changeState()
+    }
+
+    protected open fun changeState() {
+        val point = hand.calculate()
+        state = when {
+            point == Point.BLACKJACK -> BlackJack
+            point > Point.BLACKJACK -> Bust
+            else -> Hittable
+        }
+    }
+
+    private fun isPlayableState(): Boolean = state is Hittable
+
+    open fun open(): Hand = hand
+
+    fun match(dealer: Dealer): Match {
+        return state.compare(dealer.state)
+    }
 }
 
 enum class Match {
