+++ src/main/kotlin/blackjack/domain/participant/Player.kt
@@ -1,17 +1,12 @@
 package blackjack.domain.participant
 
-import blackjack.domain.card.Card
 import blackjack.domain.card.Hand
-import blackjack.domain.card.Point
 
 class Player(
-    override val name: String,
-    override val hand: Hand = Hand.empty(),
+    name: String,
+    hand: Hand = Hand.empty(),
     state: State = Hittable,
-) : Participant {
-
-    override var state: State = state
-        private set
+) : Participant(name, hand, state) {
 
     override fun isPlayable(askHit: (String) -> Boolean): Boolean {
         if (!isPlayableState()) {
@@ -29,24 +24,5 @@ class Player(
 
     private fun isPlayableState(): Boolean {
         return state is Hittable
-    }
-
-    override fun receive(card: Card) {
-        check(isPlayableState()) { "카드를 더 이상 받을 수 없는 상태입니다." }
-        hand.add(card)
-        changeState()
-    }
-
-    override fun match(dealer: Dealer): Match {
-        return state.compare(dealer.state)
-    }
-
-    private fun changeState() {
-        val point = hand.calculate()
-        state = when {
-            point == Point.BLACKJACK -> BlackJack
-            point > Point.BLACKJACK -> Bust
-            else -> Hittable
-        }
     }
 }
